requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
flask-cors
gunicorn
numpy==1.23.5
tensorflow==2.12.0
Pillow
opencv-python-headless
keras==2.12.0
pandas
scikit-learn
matplotlib